[#84] Remove the stored test data for SpanishDict
