--- zdict/tests/dictionaries/spanish/content.json
@@ -1,19 +1,0 @@
-{"explains": [["noun",
-               [["(language)",
-                 [["a. el español (M)",
-                   "The Spanish spoken in the Americas differs somewhat from European Spanish.",
-                   "El español que se habla en América es algo diferente del que se habla en Europa."],
-                  ["b. el castellano (M)",
-                   "My knowledge of Spanish accelerated after moving to Madrid. ",
-                   "Mi conocimiento del castellano aumentó después de mudarme a Madrid. "]]]]],
-              ["adjective",
-               [["(of Spanish origin)",
-                 [["a. español",
-                   "Spanish omelette is one of my favorite dishes.",
-                   "La tortilla española es uno de mis platos favoritos."]]]]],
-              ["plural noun",
-               [["(inhabitants of Spain)",
-                 [["a. los españoles (M)",
-                   "The Spanish are known for a culture of food that is delicious and diverse. ",
-                   "Los españoles son conocidos por una cultura de comida que es deliciosa y variada. "]]]]]],
- "word": "Spanish"}
